+++ src/test/kotlin/no/nav/syfo/common/tilgangskontroll/ktor/TilgangskontrollExtensionsTest.kt
@@ -10,7 +10,7 @@ import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
-import no.nav.syfo.common.tilgangskontroll.VeilederTilgangForbiddenException
+import no.nav.syfo.common.tilgangskontroll.TilgangDeniedException
 import no.nav.syfo.common.tilgangskontroll.client.TilgangskontrollClient
 import no.nav.syfo.common.util.NAV_CALL_ID_HEADER
 import no.nav.syfo.common.util.NAV_PERSONIDENT_HEADER
@@ -110,7 +110,7 @@ class TilgangskontrollExtensionsTest {
             tilgangskontrollClient.hasAccess(callId, personIdent, token)
         } returns false
 
-        assertThrows(VeilederTilgangForbiddenException::class.java) {
+        assertThrows(TilgangDeniedException::class.java) {
             runBlocking {
                 routingContext.checkPersonAndSyfoTilgang(
                     action = action,
@@ -254,7 +254,7 @@ class TilgangskontrollExtensionsTest {
             tilgangskontrollClient.hasAccess(callId, personIdent, token)
         } returns false
 
-        assertThrows(VeilederTilgangForbiddenException::class.java) {
+        assertThrows(TilgangDeniedException::class.java) {
             runBlocking {
                 routingContext.checkPersonAndSyfoTilgang(
                     action = action,